clean up and fix generate-manual.sh
- use relative symlinks (ln -r) instead of ../<category>/<icon>.svg

- use scour to optimize the symbolic icons: removes the inkscape and sodipodi clutter, as well as unused ids

- fix some icons being incorrectly processed by the line-weight regex

+++ generate-manual.sh
@@ -2,58 +2,103 @@
 #
 # Generates the arcticons freedesktop icon theme.
 #
-# Usage: ./generate.sh [STYLE [LINE_WEIGHT]]
+# Usage: see ./generate.sh --help
 #
-# STYLE:
-#   Either "white" or "black"
-#
-# LINE_WEIGHT:
-#   The relative thickness of the generated icon lines. Can be any positive
-#   number (recommended range: about 0.75 to 1.5). A value of 1 means line
-#   weight for the 48x48 is identical to the original line weight.
-#
-# This will create an `arcticons` directory in the current working dir
-# containing the icon theme and an archive of the directory as
-# `arcticons.tar.gz`.
-#
-# WARNING! If the directory already exists, its original contents will be wiped.
-#
-# The generated icons can be defined via the `mapping.txt` file. Each line
-# represents a single icon and has the following format:
-#
-# <source_icon_name>,<destination_icon_path>
-#
-# ... where <source_icon_name> is the name of the icon in the original Arcticons
-# for Android icon set and <destination_icon_path> is the path (name and,
-# optionally, preceding directories) of the generated icon in the freedesktop
-# icon theme.
+# The generated icons can be defined via the `mapping.yaml` file e.g.:
+# '''
+# bluetooth:
+#    - apps/bluetooth
+#    - devices/bluetooth
+# '''
+# will result in the icon 'bluetooth.svg' being copied to 'apps/bluetooth.svg'
+# and 'devices/bluetooth.svg' being a relative symlink to 'apps/bluetooth.svg'
 
 set -euo pipefail
+
+# bold
+b=$'\033[1m'
+# color (for commands (blue))
+c=$'\033[34m'
+# reset
+r=$'\033[0m'
+
+function help {
+
+	cat <<EOF
+${b}USAGE$r:
+    $c$0 [OPTIONS]$r
+
+${b}OPTIONS$r:
+    $c-h$r, $c--help$r
+        shows this help
+
+    $c-d$r, $c--destination$r <destination folder>
+        The generated icon pack will be written to <destination folder>.
+        If $c--remove$r is specified, an existing destination folder will be deleted.
+        The default is "arcticons".
+
+    $c-r$r, $c--remove$r
+        The destination folder will be overwritten without asking.
+
+    $c-a$r, $c--archive$r
+        Compress the resulting icon pack into an $c<destination folder>.tar.gz$r.
+        If $c--remove$r is specified, any existing file or folder with this name will be deleted.
+    
+    $c-s$r, $c--style$r
+        Color of the resulting icons. Either 'black' or 'white'.
+
+    $c-l$r, $c--line-weight$r
+        The relative thickness of the generated icon lines. Can be any positive
+        number (recommended range: about 0.75 to 1.5). A value of 1 means line
+        weight for the 48x48 is identical to the original line weight.
+EOF
+}
+
+# go to the path of the script
+SCRIPTPATH="$(
+	cd -- "$(dirname "$0")" >/dev/null 2>&1
+	pwd -P
+)"
+cd "$SCRIPTPATH"
+
+# check the mapping.yaml file to be in the right format
+# - no whitespace in icon names
+# - root is object
+# - every value in the root object is list of strings
+yq -rf validate_mapping.jq mapping.yaml || exit 1
 
 # parse the command line options
 destination=./arcticons
 archive=false
 style=white
 line_weight=1.0
+remove=false
 while [[ $# -gt 0 ]]; do
 	case $1 in
+	-h | --help)
+		help
+		exit 0
+		;;
+	-r | --remove)
+		remove=true
+		;;
 	-d | --destination)
-		dest=$2
-		shift
+		destination=$2
 		shift
 		;;
 	-a | --archive)
 		archive=true
-		shift
 		;;
 	-s | --style)
 		style=$2
-		shift
+		if [ "$style" != black ] && [ "$style" != white ]; then
+			echo "error: style has to be either 'black' or 'white'"
+			exit 1
+		fi
 		shift
 		;;
 	-l | --line-weight)
 		line_weight=$2
-		shift
 		shift
 		;;
 	*)
@@ -61,6 +106,7 @@ while [[ $# -gt 0 ]]; do
 		exit 1
 		;;
 	esac
+	shift
 done
 
 # look for required programs
@@ -70,13 +116,17 @@ if ! type yq >/dev/null || ! type jq >/dev/null; then
 fi
 
 inkscape=false
-if type inkscape >/dev/null; then
+if type inkscape && type scour; then
 	inkscape=true
 else
 	echo "Inkscape not found, not creating symbolic icons"
 fi
 
-rm -r "$destination" 2>/dev/null || true
+if $remove; then
+	rm -r "$destination" 2>/dev/null || true
+elif [ -e "$destination" ]; then
+	echo "Not deleting existing destination '$destination'. Trying to update it instead."
+fi
 
 # parse the yaml
 mapfile -d '' kvpairs < <(yq -cM --raw-output0 'to_entries[]' mapping.yaml)
@@ -84,58 +134,103 @@ for kvpair in "${kvpairs[@]}"; do
 	src=$(jq -r '.key' <<<"$kvpair")
 	mapfile -d '' dests < <(jq --raw-output0 '.value[]' <<<"$kvpair")
 	dest=${dests[0]}
-	echo "$src: ${dests[*]}"
+	printf '\n%s\n' "$c$src$r: ${dests[*]}"
 
-	dest_root="$destination/scalable"
+	scalable_root="$destination/scalable"
+	symbolic_root="$destination/symbolic"
+
+	# skip if all destination files already exist (only possible without --remove)
+	dests_paths=("${dests[@]/#/$scalable_root/}")
+	dests_paths=("${dests_paths[@]/%/.svg}")
+
+	all_files_exist=true
+	for file in "${dests_paths[@]}"; do
+		if [ ! -e "$file" ]; then
+			all_files_exist=false
+			break
+		fi
+	done
+	if $all_files_exist; then
+		echo "Skipping $src, all icons already exist."
+		continue
+	fi
+
+	# create destination directories
+	mkdir -p "$scalable_root/$(dirname "$dest")"
+
 	# copy the icons to the destination
-	mkdir -p "$dest_root/$(dirname "$dest")"
-
 	if [ -e "../icons/$style/$src.svg" ]; then
-		cp -v "../icons/$style/$src.svg" "$dest_root/$dest.svg"
+		cp -v "../icons/$style/$src.svg" "$scalable_root/$dest.svg"
 	elif [ -e "./icons_linux/$style/$src.svg" ]; then
-		cp -v "./icons_linux/$style/$src.svg" "$dest_root/$dest.svg"
+		cp -v "./icons_linux/$style/$src.svg" "$scalable_root/$dest.svg"
 	else
 		echo "Skipping '$src', icon not found"
 		continue
 	fi
 
 	# apply the line weight replacement
-	grep -v 'stroke-width' "$dest_root/$dest.svg" >/dev/null && sed -i 's/\(stroke:[^;]\+\)/\1;stroke-width:1px/g' "$dest_root/$dest.svg"
-	awk -i inplace -F 'stroke-width:|px' "{ print \$1 \"stroke-width:\" (\$2 * $line_weight) \"px\" \$3; }" "$dest_root/$dest.svg"
+	sed -i ':a;N;$!ba;s/\n//g' "$scalable_root/$dest.svg" # TODO rather fix the regex magic below
+	grep -v 'stroke-width' "$scalable_root/$dest.svg" >/dev/null && sed -i 's/\(stroke:[^;]\+\)/\1;stroke-width:1px/g' "$scalable_root/$dest.svg"
+	awk -i inplace -F 'stroke-width:|px' "{ print \$1 \"stroke-width:\" (\$2 * $line_weight) \"px\" \$3; }" "$scalable_root/$dest.svg"
 
 	# create symbolic links for duplicate icons
 	if [ ${#dests[@]} -gt 1 ]; then
 		for i in $(seq 1 $((${#dests[@]} - 1))); do
-			mkdir -p "$dest_root/$(dirname "${dests[$i]}")"
-			ln -vs "../${dests[0]}.svg" "$dest_root/${dests[$i]}.svg"
+			mkdir -p "$scalable_root/$(dirname "${dests[$i]}")"
+			rm "$scalable_root/${dests[$i]}.svg" 2>/dev/null || true
+			ln -vsr "$scalable_root/$dest.svg" "$scalable_root/${dests[$i]}.svg"
 		done
 	fi
 
-	# creating scalable
+	# create symbolic icons
 	if $inkscape; then
-		dest_root="$destination/symbolic"
-		src_root="$destination/scalable"
+		mkdir -p "$symbolic_root/$(dirname "$dest")"
+		echo "creating symbolic icon '$symbolic_root/$dest-symbolic.svg'"
+		inkscape \
+			--actions="select-all;object-stroke-to-path" \
+			--export-filename="$symbolic_root/$dest-symbolic.svg" \
+			--export-overwrite \
+			"$scalable_root/$dest.svg" || true
+		# shellcheck disable=SC2015
+		scour --quiet "$symbolic_root/$dest-symbolic.svg" "$symbolic_root/$dest-symbolic.tmp.svg" &&
+			rm "$symbolic_root/$dest-symbolic.svg" &&
+			mv "$symbolic_root/$dest-symbolic.tmp.svg" "$symbolic_root/$dest-symbolic.svg" || true
 
-		mkdir -p "$dest_root/$(dirname "$dest")"
-		inkscape --actions="select-all;object-stroke-to-path" --export-filename="$dest_root/$dest-symbolic.svg" "$src_root/$dest.svg" || true
-		#rm $src_root/$dest*.0.svg || true
+		# why is this line here?
+		rm "$scalable_root/$dest*.0.svg" 2>/dev/null || true
 
 		if [ ${#dests[@]} -gt 1 ]; then
 			for i in $(seq 1 $((${#dests[@]} - 1))); do
-				mkdir -p "$dest_root/$(dirname "${dests[$i]}")"
-				ln -vs "../${dests[0]}-symbolic.svg" "$dest_root/${dests[$i]}-symbolic.svg"
+				mkdir -p "$symbolic_root/$(dirname "${dests[$i]}")"
+				ln -vs "../${dests[0]}-symbolic.svg" "$symbolic_root/${dests[$i]}-symbolic.svg"
 			done
 		fi
 	fi
 done
 
+# link all the sizes to scalable
 folders=(8x8 16x16 16x16@2x 18x18 18x18@2x 22x22 22x22@2x 24x24 24x24@2x 32x32 32x32@2x 42x42 48x48 48x48@2x 64x64 64x64@2x 84x84 96x96 128x128)
 for folder in "${folders[@]}"; do
-	ln -sv scalable "arcticons/$folder"
+	ln -svfT scalable "$destination/$folder"
 done
 
-cp index.theme arcticons/
+# copy the index.theme
+cp index.theme "$destination/"
 
+# possibly create the archive
 if $archive; then
-	tar czf arcticons.tar.gz arcticons
+	archivedest="$destination.tar.gz"
+	if $remove; then
+		rm "$archivedest" 2>/dev/null || true
+	elif [ -e "$archivedest" ]; then
+		read -rp "archive destination '$archivedest' already exists, delete? [y/n] " -n 1 choice
+		echo # \n
+		if [ "$choice" == y ]; then
+			rm "$archivedest" 2>/dev/null || true
+		else
+			echo "Not deleting existing archive '$archivedest'!"
+			exit 0
+		fi
+	fi
+	tar czf "$archivedest" "$destination"
 fi
